requirements.txt
# Core AI/ML libraries
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Deep learning frameworks
tensorflow>=2.13.0
pytorch-lightning>=2.0.0

# Natural Language Processing
spacy>=3.6.0
nltk>=3.8.0
sentence-transformers>=2.2.0

# Data processing and analysis
datasets>=2.12.0
huggingface-hub>=0.15.0
tokenizers>=0.13.0

# Vector databases and search
faiss-cpu>=1.7.4
chromadb>=0.3.0
pinecone-client>=2.2.0

# Web and API
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
pre-commit>=3.3.0

# Logging and monitoring
loguru>=0.7.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
python-dateutil>=2.8.0